+++ cmd/down.go
@@ -52,11 +52,11 @@ func Down() *cobra.Command {
 
 			manifestOpts := contextCMD.ManifestOptions{Filename: devPath, Namespace: namespace, K8sContext: k8sContext}
 			if devPath != "" {
-				workdir := utils.GetWorkdirFromManifestPath(devPath)
+				workdir := model.GetWorkdirFromManifestPath(devPath)
 				if err := os.Chdir(workdir); err != nil {
 					return err
 				}
-				devPath = utils.GetManifestPathFromWorkdir(devPath, workdir)
+				devPath = model.GetManifestPathFromWorkdir(devPath, workdir)
 			}
 			manifest, err := contextCMD.LoadManifestWithContext(ctx, manifestOpts)
 			if err != nil {
