+++ cmd/stack/destroy.go
@@ -39,11 +39,11 @@ func Destroy(ctx context.Context) *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) error {
 			oktetoLog.Warning("'okteto stack destroy' is deprecated in favor of 'okteto destroy', and will be removed in a future version")
 			if len(stackPath) == 1 {
-				workdir := utils.GetWorkdirFromManifestPath(stackPath[0])
+				workdir := model.GetWorkdirFromManifestPath(stackPath[0])
 				if err := os.Chdir(workdir); err != nil {
 					return err
 				}
-				stackPath[0] = utils.GetManifestPathFromWorkdir(stackPath[0], workdir)
+				stackPath[0] = model.GetManifestPathFromWorkdir(stackPath[0], workdir)
 			}
 			s, err := contextCMD.LoadStackWithContext(ctx, name, namespace, stackPath)
 			if err != nil {
